+++ Code/Teensy_FC/src/04Volt_Curr/Voltmeter.h
@@ -6,13 +6,13 @@
 
 class Voltmeter {
     public:
-        Voltmeter(Drone_Data_t *drone_data ,int voltmeterPin, int currentPin , int voltmeter_calibration_factor);
+        Voltmeter(Drone_Data_t *drone_data ,int voltmeterPin, int currentPin , float voltmeter_calibration_factor);
         void read_bat_data();
     private:
         Drone_Data_t *_drone_data;
         int _voltmeterPin;
         int _currentPin;
-        int _voltmeter_calibration_factor;
+        float _voltmeter_calibration_factor;
 
 };
 #endif
